Moves the Strings extensions to a new file.
