+++ lib/screens/transaction/transaction_mixin.dart
@@ -1,6 +1,7 @@
 import 'package:provenance_wallet/common/pw_design.dart';
 import 'package:provenance_wallet/screens/transaction/transaction_field_title.dart';
 import 'package:provenance_wallet/screens/transaction/transaction_field_value.dart';
+import 'package:provenance_wallet/util/extensions/string_extensions.dart';
 import 'package:provenance_wallet/util/messages/message_field.dart';
 import 'package:provenance_wallet/util/strings.dart';
